# docs/swagger.json
{
    "schemes": [
        "http",
        "https"
    ],
    "swagger": "2.0",
    "info": {
        "description": "REST API service untuk platform crowdfunding Web3. API ini berfungsi sebagai lapisan data off-chain yang berkomunikasi dengan frontend. State krusial dan transaksi finansial ditangani oleh smart contract.",
        "title": "Web3 Crowdfunding API",
        "termsOfService": "http://swagger.io/terms/",
        "contact": {
            "name": "API Support",
            "email": "support@web3crowdfunding.com"
        },
        "license": {
            "name": "MIT",
            "url": "https://opensource.org/licenses/MIT"
        },
        "version": "1.0"
    },
    "host": "103.197.188.137:3000",
    "basePath": "/api/v1",
    "paths": {
...
        "/projects": {
            "get": {
                "description": "Retrieve list of all crowdfunding projects",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Projects"
                ],
                "summary": "Get all projects",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/model.ProjectSwagger"
                            }
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/model.ErrorResponse"
                        }
                    }
                }
            },
            "post": {
                "description": "Create a new crowdfunding project entry. UUID v7 will be auto-generated",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Projects"
                ],
                "summary": "Create new project",
                "parameters": [
                    {
                        "description": "Project data",
                        "name": "project",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/model.ProjectCreate"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created",
                        "schema": {
                            "$ref": "#/definitions/model.ProjectSwagger"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/model.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/model.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/projects/{id}": {
            "get": {
                "description": "Get detailed information about a specific project",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Projects"
                ],
                "summary": "Get project by ID",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Project ID (UUID v7)",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/model.ProjectSwagger"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/model.ErrorResponse"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/model.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/model.ErrorResponse"
                        }
                    }
                }
            },
            "patch": {
                "description": "Partially update project information",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Projects"
                ],
                "summary": "Update project",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Project ID (UUID v7)",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Fields to update",
                        "name": "project",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/model.ProjectPatch"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/model.ProjectSwagger"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/model.ErrorResponse"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/model.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/model.ErrorResponse"
                        }
                    }
                }
            }
        },
...
        "/projects/{id}/investors": {
            "get": {
                "description": "Get all investor wallet addresses for a project",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Projects"
                ],
                "summary": "Get project investors",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Project ID (UUID v7)",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/model.InvestorsResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/model.ErrorResponse"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/model.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/model.ErrorResponse"
                        }
                    }
                }
            },
            "post": {
                "description": "Add an investor's wallet address to a project",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Projects"
                ],
                "summary": "Add investor to project",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Project ID (UUID v7)",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Investor wallet address",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/model.AddInvestorRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/model.GenericMessage"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/model.ErrorResponse"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/model.ErrorResponse"
                        }
                    },
                    "409": {
                        "description": "Conflict",
                        "schema": {
                            "$ref": "#/definitions/model.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/model.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/projects/{id}/investors/{walletAddress}": {
            "delete": {
                "description": "Remove an investor's wallet address from a project",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Projects"
                ],
                "summary": "Remove investor from project",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Project ID (UUID v7)",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Investor wallet address",
                        "name": "walletAddress",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/model.GenericMessage"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/model.ErrorResponse"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/model.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/model.ErrorResponse"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
        "model.AddInvestorRequest": {
            "type": "object",
            "properties": {
                "wallet_address": {
                    "type": "string",
                    "example": "0x742d35Cc6634C0532925a3b844Bc9e7595f0bEb"
                }
            }
        },
...
        "model.ErrorResponse": {
            "type": "object",
            "properties": {
                "error": {
                    "type": "string",
                    "example": "Resource not found"
                }
            }
        },
...
        "model.GenericMessage": {
            "type": "object",
            "properties": {
                "message": {
                    "type": "string",
                    "example": "Investor added successfully"
                }
            }
        },
        "model.InvestorsResponse": {
            "type": "object",
            "properties": {
                "investors": {
                    "type": "array",
                    "items": {
                        "type": "string"
                    },
                    "example": [
                        "[\"0x111...\"",
                        " \"0x222...\"]"
                    ]
                }
            }
        },
        "model.ProjectCreate": {
            "type": "object",
            "properties": {
                "cover_image_url": {
                    "type": "string",
                    "example": "https://example.com/image.jpg"
                },
                "creator_wallet_address": {
                    "type": "string",
                    "example": "0x742d35Cc6634C0532925a3b844Bc9e7595f0bEb"
                },
                "description": {
                    "type": "string",
                    "example": "This is an amazing Web3 game"
                },
                "developer_name": {
                    "type": "string",
                    "example": "GameDev Studios"
                },
                "game_type": {
                    "type": "string",
                    "example": "web3"
                },
                "genre": {
                    "type": "string",
                    "example": "RPG"
                },
                "title": {
                    "type": "string",
                    "example": "My Awesome Game"
                }
            }
        },
        "model.ProjectPatch": {
            "type": "object",
            "properties": {
                "cover_image_url": {
                    "type": "string",
                    "example": "https://example.com/new-image.jpg"
                },
                "description": {
                    "type": "string",
                    "example": "Updated description"
                },
                "developer_name": {
                    "type": "string",
                    "example": "New Studio Name"
                },
                "game_type": {
                    "type": "string",
                    "example": "web3"
                },
                "genre": {
                    "type": "string",
                    "example": "Action RPG"
                },
                "title": {
                    "type": "string",
                    "example": "Updated Game Title"
                }
            }
        },
        "model.ProjectSwagger": {
            "type": "object",
            "properties": {
                "cover_image_url": {
                    "type": "string",
                    "example": "https://example.com/cover.jpg"
                },
                "created_at": {
                    "type": "string"
                },
                "creator_wallet_address": {
                    "type": "string",
                    "example": "0x742d35Cc6634C0532925a3b844Bc9e7595f0bEb"
                },
                "description": {
                    "type": "string",
                    "example": "An amazing blockchain RPG game"
                },
                "developer_name": {
                    "type": "string",
                    "example": "Epic Games Studio"
                },
                "game_type": {
                    "type": "string",
                    "example": "web3"
                },
                "genre": {
                    "type": "string",
                    "example": "RPG"
                },
                "id": {
                    "type": "string",
                    "example": "0199fb01-ae3c-7c26-b70a-8f221585ccb4"
                },
                "investor_wallet_addresses": {
                    "type": "array",
                    "items": {
                        "type": "string"
                    },
                    "example": [
                        "[\"0xabc...\"]"
                    ]
                },
                "title": {
                    "type": "string",
                    "example": "Epic RPG Game"
                },
                "updated_at": {
                    "type": "string"
                }
            }
        },
...
    },
...
}